refactor: Reorganize EventCalendar folder structure
- Moved EventCalendar.css and EventCalendar.jsx to src/components/EventCalendar/ folder
- Moved utils.js to src/lib/events-calendar/ folder
- Updated import paths in EventCalendar.jsx and EventCalendarPage.jsx

Addresses PR feedback for better code organization
