+++ src/lib/events-calendar/utils.js
